Project name change and Elastic Cloud added
- changed project name as it was causing issues publishing to docker hub
- added elastic cloud support with parameterised connection
