+++ .gitignore
@@ -4,3 +4,6 @@ GlobalIntegrationApi/obj
 obj
 GlobalIntegrationApi/Properties/PublishProfiles
 *.user
+bin
+BuildingBlocks
+Common
