--- GlobalIntegrationApi/Dockerfile
@@ -1,26 +1,0 @@
-#See https://aka.ms/customizecontainer to learn how to customize your debug container and how Visual Studio uses this Dockerfile to build your images for faster debugging.
-
-FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
-USER app
-WORKDIR /app
-EXPOSE 80
-#EXPOSE 8080
-#EXPOSE 8081
-
-FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
-ARG BUILD_CONFIGURATION=Release
-WORKDIR /src
-COPY ["src/Microservices/GlobalIntegrationApi/GlobalIntegrationApi.csproj", "src/Microservices/GlobalIntegrationApi/"]
-RUN dotnet restore "./src/Microservices/GlobalIntegrationApi/./GlobalIntegrationApi.csproj"
-COPY . .
-WORKDIR "/src/src/Microservices/GlobalIntegrationApi"
-RUN dotnet build "./GlobalIntegrationApi.csproj" -c $BUILD_CONFIGURATION -o /app/build
-
-FROM build AS publish
-ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "./GlobalIntegrationApi.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
-
-FROM base AS final
-WORKDIR /app
-COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "GlobalIntegrationApi.dll"]
