+++ docker-compose.override.yml
@@ -1,7 +1,7 @@
 version: '3.4'
 
 services:
-  globalintegrationapi:
+  gateway-global-integration-api:
     environment:
       - ASPNETCORE_ENVIRONMENT=Development
       - ASPNETCORE_HTTP_PORTS=8080
