+++ gateway-global-integration-api/Controllers/RestartConsumerController.cs
