+++ gateway-global-integration-api/Controllers/StopConsumerController.cs
