+++ gateway-global-integration-api/GlobalIntegrationContext.cs
