+++ gateway-global-integration-api/IntegrationEvents/EventHandling/NewRsiMessageRecievedIntegrationEventHandler.cs
