+++ gateway-global-integration-api/IntegrationEvents/EventHandling/RsiMessagePublishedIntegrationEventHandler.cs
