+++ gateway-global-integration-api/IntegrationEvents/Events/RsiMessagePublishedIntegrationEvent.cs
