+++ gateway-global-integration-api/IntegrationEvents/Events/StopConsumerRequestIntegrationEvent.cs
