+++ gateway-global-integration-api/Properties/PublishProfiles/GatewayRequestApiReg.pubxml
@@ -1,0 +1,24 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<!--
+https://go.microsoft.com/fwlink/?LinkID=208121.
+-->
+<Project>
+  <PropertyGroup>
+    <WebPublishMethod>Container</WebPublishMethod>
+    <ContainerPublishMethod>NetSdk</ContainerPublishMethod>
+    <ResourceId>/subscriptions/c3a3d8e1-1944-4ac2-ba0b-d1c3f7408db9/resourceGroups/GatewayResources/providers/Microsoft.ContainerRegistry/registries/GatewayRequestApiReg</ResourceId>
+    <ResourceName>GatewayRequestApiReg</ResourceName>
+    <ResourceGroup>GatewayResources</ResourceGroup>
+    <Subscription>Visual Studio Enterprise Subscription</Subscription>
+    <RegistryUrl>gatewayrequestapireg.azurecr.io</RegistryUrl>
+    <UserName />
+    <PublishImageTag>latest</PublishImageTag>
+    <PublishProvider>ContainerRegistry</PublishProvider>
+    <LastUsedBuildConfiguration>Release</LastUsedBuildConfiguration>
+    <LastUsedPlatform>Any CPU</LastUsedPlatform>
+    <RuntimeIdentifier>linux-x64</RuntimeIdentifier>
+    <ProjectGuid>146cf42b-f42e-497c-98b0-0104f74be65a</ProjectGuid>
+    <_IsDockerfilePresent>true</_IsDockerfilePresent>
+    <_TargetId>NetSdkAzureContainerRegistry</_TargetId>
+  </PropertyGroup>
+</Project>
