+++ gateway-global-integration-api/Properties/PublishProfiles/registry.hub.docker.com_attonbomb1.pubxml
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<!--
+https://go.microsoft.com/fwlink/?LinkID=208121.
+-->
+<Project>
+  <PropertyGroup>
+    <WebPublishMethod>Custom</WebPublishMethod>
+    <DockerPublish>true</DockerPublish>
+    <RegistryUrl>https://registry.hub.docker.com/attonbomb</RegistryUrl>
+    <UserName>attonbomb</UserName>
+    <PublishImageTag>latest</PublishImageTag>
+    <PublishProvider>ContainerRegistry</PublishProvider>
+    <LastUsedBuildConfiguration>Release</LastUsedBuildConfiguration>
+    <LastUsedPlatform>Any CPU</LastUsedPlatform>
+    <ProjectGuid>33998c27-727b-4fde-b65c-1b7072c4e666</ProjectGuid>
+    <_TargetId>DockerContainerRegistry</_TargetId>
+  </PropertyGroup>
+</Project>
