+++ gateway-global-integration-api/Properties/launchSettings.json
@@ -46,7 +46,7 @@
     "anonymousAuthentication": true,
     "iisExpress": {
       "applicationUrl": "http://localhost:61003",
-      "sslPort": 44323
+      "sslPort": 44324
     }
   }
 }
