+++ gateway-global-integration-api/Queries/GlobalRequestViewModels.cs
