+++ gateway-global-integration-api/Queries/IGlobalDataQueries.cs
