+++ gateway-global-integration-api/Services/GlobalIntegrationServices.cs
