+++ gateway-global-integration-api/Services/IGlobalIntegrationServices.cs
