+++ gateway-global-integration-api/appsettings.Development.json
@@ -22,8 +22,14 @@
       "Application": "GlobalIntegrationApi"
     }
   },
+  "ConnectionStrings": {
+    //"DefaultConnection": "Data Source=DADS-LENOVO-LAP\\SQLEXPRESS;Database=Gateway;Integrated Security=True;User ID=DADS-LENOVO-LAP\\thoma;MultipleActiveResultSets=false;TrustServerCertificate=True;",
+    "DefaultConnection": "Data Source={#host};Database={#dbName};Integrated Security=False;User ID=sa;Password={#dbPassword};MultipleActiveResultSets=false;TrustServerCertificate=True;",
+    //"DefaultConnection": "Server=tcp:global-integration-sql-server.database.windows.net,1433;Initial Catalog=Global_Integration;Persist Security Info=False;User ID=gatewayadmin;Password=password@12345#;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;",
+    "EventBus": "Endpoint=sb://gateway-messaging.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=qkBjKbAm53hXoA4h36Keg2QdRssYUOvJR+ASbIOzGfo="
+  },
   "EventBus": {
-    "ProviderName": "ServiceBus", //"RabbitMQ",
+    "ProviderName": "RabbitMQ",
     "SubscriptionClientName": "gateway_global_integration_evts",
     "Username": "guest",
     "Password": "guest",
@@ -32,5 +38,6 @@
     "RetryCount": 5,
     "GlobalIntegrationRoutingKey": "#.IntegrationEvent",
     "GlobalIntegrationQueueName": "gateway_global_integration_evts"
-  }
+  },
+  "AllowedHosts": "*"
 }
