+++ gateway-global-integration-api/appsettings.json
@@ -34,6 +34,11 @@
     "DefaultConnection": "Server=tcp:global-integration-sql-server.database.windows.net,1433;Initial Catalog=Global_Integration;Persist Security Info=False;User ID=gatewayadmin;Password=password@12345#;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;",
     "EventBus": "Endpoint=sb://gateway-messaging.servicebus.windows.net/;SharedAccessKeyName=RootManageSharedAccessKey;SharedAccessKey=qkBjKbAm53hXoA4h36Keg2QdRssYUOvJR+ASbIOzGfo="
   },
+  "ElasticCloud": {
+    "CloudId": "My_deployment:dWtzb3V0aC5henVyZS5lbGFzdGljLWNsb3VkLmNvbTo0NDMkMDMzNzVjZDdkMzllNDE0ODgwZjc2NjU4NjhmYzlmYTEkNzlmYThlOGU5YjcwNDY2YWFlMzM3M2JhYTg2ZDhkZDc=",
+    "CloudUser": "elastic",
+    "CloudPass": "QLI3OVBRpBQikrBbK09JITRQ"
+  },
   "EventBus": {
     "ProviderName": "ServiceBus", //"RabbitMQ",
     "SubscriptionClientName": "gateway_global_integration_evts",
